requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
future==1.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
line-bot-sdk==3.19.1
MarkupSafe==3.0.2
multidict==6.6.4
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.32.5
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
